+++ tests/packerlicious/test_builder_virtualbox.py
@@ -3,17 +3,27 @@ import pytest
 import packerlicious.builder as builder
 
 
-class TestVMwareIsoBuilder(object):
+class TestVirtualBoxOvfBuilder(object):
 
     def test_required_fields_missing(self):
-        b = builder.VMwareIso()
+        b = builder.VirtualboxOvf()
+
+        with pytest.raises(ValueError) as excinfo:
+            b.to_dict()
+        assert 'required' in str(excinfo.value)
+
+
+class TestVirtualBoxIsoBuilder(object):
+
+    def test_required_fields_missing(self):
+        b = builder.VirtualboxIso()
 
         with pytest.raises(ValueError) as excinfo:
             b.to_dict()
         assert 'required' in str(excinfo.value)
 
     def test_is_checksum_mutually_exclusive(self):
-        b = builder.VMwareIso(
+        b = builder.VirtualboxIso(
             iso_url="/url/to/iso",
             iso_checksum_type=builder.VirtualboxIso.MD5,
             iso_checksum="my_checksum",
@@ -22,15 +32,4 @@ class TestVMwareIsoBuilder(object):
 
         with pytest.raises(ValueError) as excinfo:
             b.to_dict()
-        assert 'VMwareIso: only one of the following can be specified: iso_checksum, iso_checksum_url' == str(
-            excinfo.value)
-
-
-class TestVMwareVmxBuilder(object):
-
-    def test_required_fields_missing(self):
-        b = builder.VMwareVmx()
-
-        with pytest.raises(ValueError) as excinfo:
-            b.to_dict()
-        assert 'required' in str(excinfo.value)
+        assert 'VirtualboxIso: only one of the following can be specified: iso_checksum, iso_checksum_url' == str(excinfo.value)