+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 
 ### FEATURES:
+* builder/VirtualboxIso: add support for packer's virtualbox iso builder
+* builder/VirtualboxOvf: add support for packer's virtualbox ovf builder
 * builder/VMwareIso: add support for packer's vmware iso builder
 * builder/VMwareVmx: add support for packer's vmware vmx builder